requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.8
django-crispy==1.8.1
django-crispy-forms==1.13.0
Pillow==8.3.2
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
